refactor (test/facade): Ts'ify test/facade
Translate AtScript in test/facade to TypeScript
